Remove all hardcoded Descope URLs and project IDs
- Create shared config.ts module with constants and configuration loading
- Update create-server.ts to load dynamic configuration from Netlify Blobs
- Replace all hardcoded api.descope.com references with constants
- Update UI JavaScript to use dynamic constants
- Make HTML input default values dynamic
- Ensure all endpoints use configuration from Netlify Blobs with env fallback

Now the entire application uses dynamic configuration without any hardcoded values,
allowing full customization of Descope base URL and project ID through the UI.
